refactor(parser): Move splitTopLevelStatements helper to shared.ts
