+++ packages/@birdcc/parser/src/declarations/protocol-radv.ts
@@ -1,4 +1,5 @@
 import type { ProtocolStatement, SourceRange } from "../types.js";
+import { splitTopLevelStatements } from "./shared.js";
 
 type TokenRange = (token: string) => SourceRange;
 
@@ -19,37 +20,6 @@ const parseBoolToken = (value: string | undefined): boolean | undefined => {
   }
 
   return undefined;
-};
-
-const splitTopLevelStatements = (body: string): string[] => {
-  const statements: string[] = [];
-  let depth = 0;
-  let start = 0;
-
-  for (let index = 0; index < body.length; index += 1) {
-    const char = body[index];
-    if (char === "{") {
-      depth += 1;
-      continue;
-    }
-
-    if (char === "}") {
-      depth = Math.max(0, depth - 1);
-      continue;
-    }
-
-    if (char === ";" && depth === 0) {
-      statements.push(body.slice(start, index));
-      start = index + 1;
-    }
-  }
-
-  const tail = body.slice(start).trim();
-  if (tail.length > 0) {
-    statements.push(tail);
-  }
-
-  return statements;
 };
 
 const parseRadvCustomOptionParts = (
