+++ packages/@birdcc/parser/src/declarations/shared.ts
@@ -66,6 +66,37 @@ export const PROTOCOL_STATEMENT_TYPES = new Set([
   "expression_statement",
 ]);
 
+export const splitTopLevelStatements = (body: string): string[] => {
+  const statements: string[] = [];
+  let depth = 0;
+  let start = 0;
+
+  for (let index = 0; index < body.length; index += 1) {
+    const char = body[index];
+    if (char === "{") {
+      depth += 1;
+      continue;
+    }
+
+    if (char === "}") {
+      depth = Math.max(0, depth - 1);
+      continue;
+    }
+
+    if (char === ";" && depth === 0) {
+      statements.push(body.slice(start, index));
+      start = index + 1;
+    }
+  }
+
+  const tail = body.slice(start).trim();
+  if (tail.length > 0) {
+    statements.push(tail);
+  }
+
+  return statements;
+};
+
 export const TABLE_TYPES = new Set([
   "routing",
   "ipv4",
